# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Browser Calculator API",
    "description": "Evaluate a single arithmetic expression and accept one form submission.",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {}
  },
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "health",
        "summary": "Report that the service is up.",
        "tags": [
          "health"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluate": {
      "post": {
        "operationId": "evaluateExpression",
        "summary": "Evaluate an arithmetic expression.",
        "tags": [
          "calculator"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "expression"
                ],
                "additionalProperties": false,
                "properties": {
                  "expression": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "expression",
                    "result"
                  ],
                  "properties": {
                    "expression": {
                      "type": "string"
                    },
                    "result": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/form": {
      "post": {
        "operationId": "submitForm",
        "summary": "Validate a contact form submission and acknowledge it.",
        "tags": [
          "tools"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "message"
                ],
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "email": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 254,
                    "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$"
                  },
                  "message": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "name",
                    "email",
                    "message",
                    "acknowledged",
                    "receivedAt"
                  ],
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "acknowledged": {
                      "type": "boolean"
                    },
                    "receivedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "calculator",
      "description": "Evaluate arithmetic."
    },
    {
      "name": "tools",
      "description": "Small stateless utilities."
    },
    {
      "name": "health",
      "description": "Liveness."
    }
  ]
}